Wrote an OAuth Provider for use in QML
Wrote an OAuth provider for use as a base class for web-based services.
Currently implemented a Twitter subclass, but it should be easily
extensible to others. Ultimate goal is to incorporate it as a QML
plugin, rather than a context property. Currently allows the user to
login to their account. Still lots to do.