Removed the circular class references
I kicked out the child classes into top-level, and removed the circular
references. Someone smarter than I can likely figure out a better way to
accomplish what I was trying to do.

However, this code causes a segmentation fault when the QML is trying to
populate the ListModel with tweets. Need to investigate.